+++ engine/runtime-integration-tests/src/test/java/org/enso/interpreter/test/SignatureTest.java
@@ -1027,7 +1027,20 @@ public class SignatureTest extends ContextTest {
   }
 
   @Test
-  public void returnTypeCheckByLastStatementOfMain() throws Exception {
+  public void returnTypeCheckByLastStatementOfMainTextFirst() throws Exception {
+    var main = assertTypeCheckByLastStatementOfMain("Text & Integer");
+    assertTrue(main.isString());
+    assertEquals("42", main.asString());
+  }
+
+  @Test
+  public void returnTypeCheckByLastStatementOfMainIntFirst() throws Exception {
+    var main = assertTypeCheckByLastStatementOfMain("Integer & Text");
+    assertTrue(main.fitsInInt());
+    assertEquals(42, main.asInt());
+  }
+
+  private Value assertTypeCheckByLastStatementOfMain(String cast) throws Exception {
     final URI uri = new URI("memory://rts.enso");
     final Source src =
         Source.newBuilder(
@@ -1036,18 +1049,17 @@ public class SignatureTest extends ContextTest {
                 from Standard.Base import all
 
                 fn =
-                    (42 : Text & Integer)
+                    (42 : ${cast})
 
                 Text.from (that:Integer) = that.to_text
-                """,
+                """
+                    .replace("${cast}", cast),
                 uri.getAuthority())
             .uri(uri)
             .buildLiteral();
 
     var module = ctx.eval(src);
-    var main = module.invokeMember(MethodNames.Module.EVAL_EXPRESSION, "fn");
-    assertEquals(42, main.asInt());
-    assertEquals("42", main.asString());
+    return module.invokeMember(MethodNames.Module.EVAL_EXPRESSION, "fn");
   }
 
   /**