+++ engine/runtime/src/main/java/org/enso/interpreter/node/typecheck/AllOfTypesCheckNode.java
@@ -40,11 +40,19 @@ final class AllOfTypesCheckNode extends AbstractTypeCheckNode {
       if (result == null) {
         return null;
       }
-      values[at] = result;
       valueTypes[at] = types.getType(result);
+      if (result instanceof EnsoMultiValue emv) {
+        result =
+            EnsoMultiValue.CastToNode.getUncached()
+                .findTypeOrNull(valueTypes[at], emv, false, true);
+      }
+      if (result == null) {
+        return null;
+      }
+      values[at] = result;
       at++;
     }
-    return EnsoMultiValue.create(valueTypes, values);
+    return EnsoMultiValue.create(valueTypes, valueTypes.length, values);
   }
 
   @Override